working on en passant
added a piece super class to pawn class to share utility moves between pieces. also getting closer to en passant... it has to be added after the last pawn move (update_pieces).

next steps is to process the pgn notation for that move if it's taken from the attacks array for the pawn that has that option.

also, if the option is not taken, we need to make sure en passant is not available on subsequent moves.